requirements.txt
qdrant-client 
langchain
langchain_community
langchain_ollama
psycopg2-binary 
sqlalchemy
langchain-qdrant
torch 
torchvision 
pillow
opencv-python 
sentence-transformers
python-dotenv
# dbt-postgres
# vanna
# 'vanna[qdrant,ollama,postgres]'
langgraph
fastapi
uvicorn
unstructured
streamlit
langchain-nvidia-ai-endpoints